fixed padding issue

+++ S15/transformer_lightning/dataset.py
@@ -47,11 +47,11 @@ class BilingualDataset(Dataset):
         self.tgt_lang = tgt_lang
 
         self.sos_token = torch.tensor(
-            [tokenizer_tgt.token_to_id("[SOS]")], dtype=torch.int64)
+            [tokenizer_tgt.token_to_id("[SOS]")], dtype=torch.int8)
         self.eos_token = torch.tensor(
-            [tokenizer_tgt.token_to_id("[EOS]")], dtype=torch.int64)
+            [tokenizer_tgt.token_to_id("[EOS]")], dtype=torch.int8)
         self.pad_token = torch.tensor(
-            [tokenizer_tgt.token_to_id("[PAD]")], dtype=torch.int64)
+            [tokenizer_tgt.token_to_id("[PAD]")], dtype=torch.int8)
 
     def __len__(self):
         return len(self.ds)
